server/requirements.txt
# One-click runtime deps for this repo's server scripts.
#
# Notes:
# - `sam2` is installed from the official GitHub repo via a VCS URL.
# - SAM2 recommends torch>=2.5.1 / torchvision>=0.20.1 (per upstream README).
# - On macOS / CPU-only machines, we set SAM2_BUILD_CUDA=0 during install to avoid CUDA builds.

numpy>=1.24
Pillow>=10.0

torch>=2.5.1
torchvision>=0.20.1

# SAM2 code (provides `import sam2`)
git+https://github.com/facebookresearch/sam2.git

# HTTP server
fastapi>=0.110
uvicorn>=0.29
python-multipart>=0.0.9
